separated tests out into a separate file from main mailroom
